refactor: 팔로우, 언팔로우 API 리팩토링
- FollowService 메서드에 주체 회원의 id만 넘겨줌
- 언팔로우를 할 때 팔로우 id를 통해 삭제함
- 팔로우 API 응답을 FollowDto로 변경 (MemberDto 추가)
- Follow 엔티티 컬럼 이름 직관적으로 수정
- 전반적인 API의 요청과 응답의 값을 Camel Case로 수정

+++ src/main/java/com/youngxpepp/instagramcloneserver/domain/follow/service/FollowService.java
@@ -1,15 +1,10 @@
 package com.youngxpepp.instagramcloneserver.domain.follow.service;
 
-import javax.validation.Valid;
-
 import lombok.AllArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.youngxpepp.instagramcloneserver.domain.follow.dto.FollowRequestDto;
-import com.youngxpepp.instagramcloneserver.domain.follow.dto.UnfollowRequestDto;
+import com.youngxpepp.instagramcloneserver.domain.follow.dto.FollowDto;
 import com.youngxpepp.instagramcloneserver.domain.follow.model.Follow;
 import com.youngxpepp.instagramcloneserver.domain.follow.repository.FollowRepository;
 import com.youngxpepp.instagramcloneserver.domain.member.model.Member;
@@ -19,39 +14,41 @@ import com.youngxpepp.instagramcloneserver.global.error.exception.BusinessExcept
 
 @Service
 @AllArgsConstructor
-@Slf4j
 public class FollowService {
 
 	private FollowRepository followRepository;
 	private MemberRepository memberRepository;
 
 	@Transactional
-	public void follow(Long memberId, FollowRequestDto dto) throws BusinessException {
-		Member fromMember = memberRepository.getOne(memberId);
-		Member toMember = memberRepository.findByNickname(dto.getMemberNickname())
+	public FollowDto follow(Long followingMemberId, Long followedMemberId) {
+		Member followingMember = memberRepository.findById(followingMemberId)
+			.orElseThrow(() -> new BusinessException(ErrorCode.ENTITY_NOT_FOUND));
+		Member followedMember = memberRepository.findById(followedMemberId)
 			.orElseThrow(() -> new BusinessException(ErrorCode.ENTITY_NOT_FOUND));
 
-		followRepository.findByFromMemberIdAndToMemberId(fromMember.getId(), toMember.getId())
+		followRepository.findByFollowingMemberIdAndFollowedMemberId(followingMember.getId(), followedMember.getId())
 			.ifPresent(follow -> {
 				throw new BusinessException(ErrorCode.ENTITY_ALREADY_EXIST);
 			});
 
 		Follow follow = Follow.builder()
-			.fromMember(fromMember)
-			.toMember(toMember)
+			.followingMember(followingMember)
+			.followedMember(followedMember)
 			.build();
 
 		followRepository.save(follow);
+
+		return FollowDto.ofFollow(follow);
 	}
 
 	@Transactional
-	public void unfollow(Long memberId, @Valid UnfollowRequestDto dto) {
-		Member fromMember = memberRepository.getOne(memberId);
-		Member toMember = memberRepository.findByNickname(dto.getMemberNickname())
+	public void unfollow(Long followingMemberId, Long followId) {
+		Follow follow = followRepository.findById(followId)
 			.orElseThrow(() -> new BusinessException(ErrorCode.ENTITY_NOT_FOUND));
 
-		Follow follow = followRepository.findByFromMemberIdAndToMemberId(fromMember.getId(), toMember.getId())
-			.orElseThrow(() -> new BusinessException(ErrorCode.ENTITY_NOT_FOUND));
+		if (!follow.getFollowingMember().getId().equals(followingMemberId)) {
+			throw new BusinessException(ErrorCode.HANDLE_ACCESS_DENIED);
+		}
 
 		followRepository.delete(follow);
 	}
